Reverse the polarity of the neutron flow.
The perspective projection that the camera-arrays
produce does indeed look weird in an equirectangular
360/vr video mapping.

But the correction warping that I'd built based on
the Unity/Google paper was doing the job backwards I think?

I only thought it was working because it also had
a bug that swapped the eyes over, which cancelled out
the bug where the eyes rendered in the wrong order.

So I had this bad projection I couldn't even focus on,
found that the equirectangular projection fixed it,
and assumed all was well.

It turned out not to be well and everything was too
tall and stretched.

Investigation fixed first that eye-switch bug which
made me realize how much better no-fix-projection-at-all
was, and then on to realizing somehow the math must
be backwards.

I remain unconvinced this is entirely right,
but it's also not even possible as far as I can figure,
so maybe this is just the best that can be done.

I welcome pull-requests from mathematicians who
have any idea how any of this works and can
fix the mess in the maths.

Especially if you also know how to make it hardware
efficient and run on the GPU etc.